requirements.txt
google-genai==1.65.0
numpy==2.4.2
python-dotenv==1.1.1
sounddevice==0.5.5
streamlit==1.54.0